Day3/requirements.txt
IMDbPY==6.9.dev20190924
lxml==4.4.1
SQLAlchemy==1.3.8
XlsxWriter==1.2.1
